proto: update unreserved characters
Remove . from the list of allowed, unreserved characters for
attributes.
This constraint is set to avoid having invalid characters in
Tom's URLs.

In the DNS-style encoding of the TomID, the . is used as separator:

        ${name}.${namespace}.${entity}.tom

If it is allowed to use . in names and namespaces, then the following
two records can be created:

        name=foo.bar in namespace=foo
        name=foo     in namespace=bar.foo

The resulting TomIDs would no longer be unique:

        foo.bar.foo.${entity}.tom
        foo.bar.foo.${entity}.tom