tom: format client raised error messages
For server-generated errors, the output has always been a JSON
answer on StdOut with exitcode 0.
For client-generated errors, the output was unformatted plaintext
on StdErr with exitcode 1.

Update to wrap the client error into a protocol result and output
the JSON for local client errors as well. Use exitcode 0 in this
case as well.

If formatting or rendering of the error fails, output all errors
on StdErr and exitcode with exit 2.