DB schema update: all.20220204001
The database schema enforces timestamps to be in UTC via constraints
checking. For timestamps created via trigger or default column value,
the timezone used is not the timezone of the session, but the default
timezone of the database itself.
This update forces all trigger default timestamps to be in timezone UTC.